Removing old style CryptoJobHandler and CryptoWorker
