--- ambry-router/src/main/java/com.github.ambry.router/CryptoJobHandler.java
@@ -1,85 +1,0 @@
-/*
- * Copyright 2017 LinkedIn Corp. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- */
-package com.github.ambry.router;
-
-import com.github.ambry.utils.Utils;
-import java.io.Closeable;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.atomic.AtomicBoolean;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-
-/**
- * Responsible for handling crypto jobs. {@link Router} instantiates this {@link CryptoJobHandler} and gives a
- * reference of the same to PutManager, GetBlobOperation and GetBlobInfoOperation which will add jobs via
- * {@link #submitJob(CryptoJob)}. On close, all pending jobs will be processed(either successfully or exception is set)
- * and any new jobs submitted after close will be ignored.
- */
-class CryptoJobHandler implements Closeable {
-  private final BlockingQueue<CryptoJob> jobQueue;
-  private final List<CryptoWorker> workerThreads;
-  private final AtomicBoolean enabled = new AtomicBoolean(false);
-
-  private static final Logger logger = LoggerFactory.getLogger(CryptoJobHandler.class);
-
-  /**
-   * Instantiates {@link CryptoJobHandler} with the job queue, {@link CryptoService} and {@link KeyManagementService}
-   * @param cryptoService the {@link CryptoService} to use to encrypt or decrypt
-   * @param kms the {@link KeyManagementService} to fetch keys from
-   */
-  CryptoJobHandler(CryptoService cryptoService, KeyManagementService kms, int workerCount) {
-    this.jobQueue = new LinkedBlockingQueue<>();
-    this.workerThreads = new ArrayList<>();
-    for (int i = 0; i < workerCount; i++) {
-      workerThreads.add(new CryptoWorker(jobQueue, cryptoService, kms, "CryptoWorker_" + i));
-    }
-  }
-
-  /**
-   * Starts up {@link CryptoJobHandler} by spinning up multiple {@link CryptoWorker} threads to assist in executing
-   * {@link CryptoJob}s submitted to {@link CryptoJobHandler} via {@link #submitJob(CryptoJob)}
-   */
-  void start() {
-    enabled.set(true);
-    for (CryptoWorker worker : workerThreads) {
-      Utils.newThread(worker, true).start();
-    }
-  }
-
-  /**
-   * Submits new job to the {@link CryptoJobHandler}
-   * @param cryptoJob the {@link CryptoJob} that needs to be executed
-   */
-  void submitJob(CryptoJob cryptoJob) {
-    if (enabled.get()) {
-      jobQueue.add(cryptoJob);
-    }
-  }
-
-  /**
-   * Shuts down the job handler. Any new jobs submitted after close are ignored. All the pending jobs will get processed
-   * for sure (either exception will be set or will get processed in usual manner)
-   */
-  public void close() {
-    if (enabled.compareAndSet(true, false)) {
-      for (CryptoWorker worker : workerThreads) {
-        worker.close();
-      }
-    }
-  }
-}
